pyproject.toml
[tool.poetry]
name = "teslab"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.87.0"
uvicorn = "^0.19.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
yandexcloud = "^0.190.0"
SQLAlchemy = "^1.4.44"
alembic = "^1.8.1"
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]
devtools = "^0.8.0"
black = "^22.6.0"
isort = "^5.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
